backend/requirements_simple.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
requests
python-dotenv
google-api-python-client
google-generativeai
httpx